use conditional requirements to ensure "chardet" listed as requirement on Python 3 (fixes #213)
Previously "chardet" was added only added when setup.py was run with Python 3.
However wheels contain a static list of requirements and a wheel-based
install will never execute setup.py at installation time.

pdfminer.six uses universal wheels for Python 2 and Python 3 so the
requirements will always be wrong on one version (see #213).

The solution is to use conditional requirements as specified in PEP 496
which are evaluated at installation time.