Handle XREFs with missing startxref after trailer
I have some PDF documents that look like:

  %PDF

  %objects
  xref
  %table
  trailer
  <</Root ..>

  %objects
  xref
  %table
  trailer
  <</Root /Prev ..>

  %objects
  xref
  %table
  trailer
  <</Root /Prev ..>

  %objects
  xref
  %table
  trailer
  <</Root /Prev ..>
  startxref
  %pointer
  %%EOF

So it looks kind of like its incrementally saved buts its missing
'startxref' and EOF after each table. Also, the early xref objects
reference objects later in the file so it is not due to incrementally
saving. The previous code failed to parse the <</Root>> dictionary
because it assumes it is terminated with 'startxref'. So the last
two xref tables would be parsed but the first two tables would not
be found. The Fallback XREF parser would then find the first XREF
table, but it would stop after hitting the 'trailer' keyword and
would not find the second XREF table.

This changes nextobject() method to allow terminating the dict
after the dict has finished parsing and not waiting for another
token like 'trailer' to indicate the production has finished. Then
when we try and parse the <</Root>> dict we use this option.

This broken PDF seems to open correctly in all the PDF viewers
I tried. I suspect because they parse the <</Root>> dict as a simple
dict production and not dict followed by 'startxref'.