Fix Memory Leak (Issue 580)
PDF parsing leaked memory. If you bracketed a parsing operation with
gc.set_debug(gc.DEBUG_SAVEALL) and gc.garbage you would see a large number
of PdfMiner that are unable to be garbage collected.

The PSBaseParser kept track of its parse state by having a _parse1 instance
variable to which it assigned member functions of the same instance. This lead
to circular references which caused the memory leaks.

This fix replaces _parse1 with a _parse_type variable that takes on a value
of the ParseType enumerable. A parse_token member function dispatches the parse
operations with an if statement. With this change, memory no longer leaks.