Fix Memory Leak with Weak References
PDF parsing leaks memory. If you surround a parsing operation with
gc.set_debug(gc.DEBUG_SAVEALL) and gc.garbage you would see a large
number of PdfMiner objects that are unable to be garbage collected.

The PSBaseParser keeps track of its parse state by having a _parse1
instance variable assigned to member functions of the same instance.
This lead to circular references which causes the memory leaks.

The fix is to change all the _parse1 assignments to be
weakref.WeakMethod objects.